Gong

+++ ClientSystem/src/main/java/presentation/guestUI/controller/HotelSearchController.java
@@ -18,7 +18,6 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.CheckBox;
-import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.DatePicker;
 import javafx.scene.control.Label;
@@ -27,6 +26,7 @@ import javafx.scene.control.TableView;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.Pane;
+import javafx.scene.layout.StackPane;
 import presentation.hotelWorkerUI.controller.HotelTable;
 import presentation.hotelWorkerUI.controller.OrderTable;
 import presentation.hotelWorkerUI.controller.TypeTable;
@@ -41,7 +41,6 @@ import vo.OrderGeneralVO;
 import vo.OrderVO;
 import vo.RoomInfoVO;
 import vo.SearchCriteriaVO;
-import javafx.scene.layout.StackPane;
 /**
  * @author 61990
  * @控制酒店预定界面
@@ -86,9 +85,8 @@ public class HotelSearchController {
 	private ComboBox<Double> minScoreInput,maxScoreInput;
 	/**
 	 * @author 61990
-	 * @lastChangedBy Byron Dong
-	 * @updateTime 2016/12/7 
-	 * @构造函数，初始化各种COMBOBOX
+	 * @lastChangedBy Harvey
+	 * @updateTime 2016/12/8 
 	 */
 	@FXML
 	private void initialize() {
@@ -503,10 +501,8 @@ public class HotelSearchController {
 		}
 		if(minLevelInput.getValue()!=null){
 			criteria.add(SearchCriteriaType.LEVEL_SPAN);
-			//TODO gy注意：将星级选择的getValue改为int；
-			//TODO gcm改了
-			vo.minLevel = minLevelInput.getValue().intValue();
-			vo.maxLevel =  maxLevelInput.getValue().intValue();
+			vo.minLevel = minLevelInput.getValue();
+			vo.maxLevel =  maxLevelInput.getValue();
 		}
 		if(minpriceInput.getText()!=null){
 			criteria.add(SearchCriteriaType.ORGIN_PRICE_SPAN);
@@ -522,9 +518,11 @@ public class HotelSearchController {
 			criteria.add(SearchCriteriaType.REMAIN_ROOM_NUM);
 			vo.remainRoomNum = Integer.parseInt(roomInput.getText());
 		}
-		//TODO gy注意：能不能把房间类型列表的选择做成复选框checkBox或者选择框choiceBox，可能ifelse会少一点儿
-//		TODO gcm 不就是复选框checkBox吗可以做成只能单选的，更方便可能，但是每次搜索一个类型的
+		
+//		TODO gcm 不就是复选框checkBox吗
+		//TODO gy 我不能用一个for循环去取到已经被勾选的框，只能单个取，给我一种可以通过循环取的方案
 //		vo.roomTypes = new ArrayList<RoomType>();
+		
 		criteria.add(SearchCriteriaType.NULL);
 		
 	}
